Moved navlink component to separate folder
