+++ test/components/t_navlink.js
@@ -3,7 +3,7 @@ import expect from 'expect';
 import { shallow } from 'enzyme';
 import { Link } from 'react-router';
 
-import NavLink from '../../src/components/navlink.js';
+import NavLink from '../../src/components/navlink/navlink.js';
 
 describe('<NavLink /> component', () => {
